requirements.txt
python-telegram-bot[job-queue]>=20.0
beautifulsoup4==4.10.0
requests==2.26.0